move text editor to wrapper component
