+++ app/canvas/features/wrapper-component/text-editor/config.ts
