+++ app/canvas/features/wrapper-component/text-editor/content-editable.tsx
