+++ app/canvas/features/wrapper-component/text-editor/example-text-editor.tsx
@@ -13,7 +13,7 @@ import { OnChangePlugin } from "./plugins/plugin-on-change";
 import { InstancePlugin } from "./plugins/plugin-instance";
 import { TreeViewPlugin } from "./plugins/tree-view-plugin";
 import { config } from "./config";
-import { createInstance } from "../tree-utils";
+import { createInstance } from "../../../../shared/tree-utils";
 
 type ExampleTextEditorProps = {
   onChange: (state: ChildrenUpdates) => void;
