+++ app/canvas/features/wrapper-component/text-editor/index.ts
