+++ app/canvas/features/wrapper-component/text-editor/lexical.ts
