+++ app/canvas/features/wrapper-component/text-editor/plugins/plugin-instance.tsx
