+++ app/canvas/features/wrapper-component/text-editor/plugins/plugin-on-change.tsx
