+++ app/canvas/features/wrapper-component/text-editor/plugins/plugin-toolbar-connector.tsx
