+++ app/canvas/features/wrapper-component/text-editor/styles.css
