+++ app/canvas/features/wrapper-component/text-editor/utils/to-lexical-nodes.tsx
