+++ app/canvas/features/wrapper-component/text-editor/utils/to-updates.ts
